Add `conda_lockfiles.types`
